chore: separate letter viewer and template editor components
