+++ frontend/src/features/generate/hooks/tinymce.hooks.tsx
